\name{HMMEMundir-package}
\alias{HMMEMundir-package}
\alias{HMMEMundir}
\docType{package}
\title{
\packageTitle{HMMEMundir}
}
\description{
\packageDescription{HMMEMundir}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HMMEMundir}
\packageIndices{HMMEMundir}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HMMEMundir}

Maintainer: \packageMaintainer{HMMEMundir}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
